Great refactoring of the project has been done. Sorting files into modules.
