--- src/app/login-input/login-input.component.ts
@@ -1,64 +1,0 @@
-import { HttpClient } from '@angular/common/http';
-import { Component, OnInit } from '@angular/core';
-import { NgForm, FormControl, FormGroup, Validators } from '@angular/forms';
-import { MatDialog } from '@angular/material/dialog';
-import { LoginInfoComponent } from '../login-info/login-info.component';
-import { LoginRequest } from '../model/LoginRequest';
-import { AuthenticationService } from '../service/authentication.service';
-import { logged, firstName, lastName } from 'src/app/app.component';
-import { CurrentUserService } from '../service/current-user.service';
-import { UserResponseDto } from '../model/UserResponseDto';
-
-@Component({
-    selector: 'app-login-input',
-    templateUrl: './login-input.component.html',
-    styleUrls: ['./login-input.component.css']
-})
-
-export class LoginInputComponent{
-    constructor(private authService: AuthenticationService,
-                public dialog: MatDialog
-                ) {}
-
-    loginControl: FormGroup = new FormGroup({
-        username: new FormControl('', [Validators.required, Validators.min(3)]),
-        password: new FormControl('', [Validators.required, Validators.min(3)])
-    });
-
-    info: string;
-    success: boolean;
-    loginValue = 'Enter login';
-    passwordValue = 'Enter passoword';
-    hide = true;
-
-    loginRequest: LoginRequest;
-
-  // tslint:disable-next-line:use-lifecycle-interface
-    ngOnInit(): void {}
-
-  // tslint:disable-next-line:typedef
-    onSubmit() {
-        this.loginRequest = this.loginControl.value;
-        this.authService.loginUser(this.loginRequest)
-        .subscribe(msg => {
-          console.log(msg.message);
-          this.info = "Pomyślnie zalogowano";
-          this.success = true;
-          
-        }, error => {
-          console.log(error.message);
-          this.info = "Nie udało się zalogować";
-          this.success = false;
-        });
-        setTimeout(() => {this.openDialog();},
-        1000);
-        
-    }
-
-
-    openDialog(): void{
-      const dialogRef = this.dialog.open(LoginInfoComponent, {data: {message:this.info, success:this.success},width: '250px'});
-    }
-
-}
-
