+++ src/app/login/component/login-info/login-info.component.html
@@ -3,7 +3,7 @@
 </div>
 <pre></pre>
 <div *ngIf="data.success; else elseBlock" class="button-form">
-    <button mat-raised-button mat-dialog-close [routerLink]="['/']">Ok</button>
+    <button mat-raised-button mat-dialog-close [routerLink]="['/']" (click)="onOkClick()">Ok</button>
 </div>
 <ng-template #elseBlock class="button-form">
     <button mat-raised-button mat-dialog-close>Ok</button>
