+++ src/app/login/component/login-input/login-input.component.html
@@ -1,0 +1,35 @@
+<mat-card class="card">
+
+	<mat-card-content>
+		<mat-form-field appearance="fill" 
+			class="input-form-field" 
+			[formGroup]="loginControl">
+			<mat-label> Nazwa użytkownika </mat-label>
+			<input matInput placeholder="Nazwa użytkownika" #login required minlength="4" formControlName="username">
+			<button mat-button matSuffix mat-icon-button aria-label="Clear" (click)="login.value=''">
+				<mat-icon>close</mat-icon>
+			</button>
+		</mat-form-field>
+	</mat-card-content>
+
+
+	<mat-card-content>
+		<mat-form-field appearance="fill" 
+			hintLabel="Max 30 znaków" 
+			class="input-form-field"
+			[formGroup]="loginControl">
+			<mat-label>Hasło</mat-label>
+			<input matInput [type]="hide ? 'password' : 'text'" formControlName="password" required placeholder="Hasło" #password minlength="6" maxlength="30">
+			<mat-icon matSuffix (click)="hide = !hide">{{hide ? 'visibility_off' : 'visibility'}}</mat-icon>
+			<mat-hint align="end">{{password.value?.length || 0}}/30</mat-hint>
+		</mat-form-field>
+	</mat-card-content>
+
+
+	<mat-card-actions>
+		<button mat-stroked-button class="input-form-field" (click)="onSubmit()">
+			Zaloguj
+		</button>
+	</mat-card-actions>
+
+</mat-card>
