+++ src/app/registration/component/registation-user/registration.component.css
