+++ src/app/registration/component/registation-user/registration.component.html
@@ -1,19 +1,19 @@
 <mat-card class="card">
     <mat-card-content>
-        <mat-form-field appearance="fill" [floatLabel]="usernameValue" class="input-form-field" hintLabel="Min. 6 znaków" [formGroup]="registrationFormGroup">
+        <mat-form-field appearance="fill" class="input-form-field" hintLabel="Min. 6 znaków" [formGroup]="registrationFormGroup">
             <mat-label> Nazwa użytkownika </mat-label>
             <input matInput placeholder="Nazwa użytkownika" #username required minlength="6" formControlName="username">
-            <button mat-button *ngIf="usernameValue" matSuffix mat-icon-button aria-label="Clear" (click)="username.value=''">
+            <button mat-button matSuffix mat-icon-button aria-label="Clear" (click)="username.value=''">
                 <mat-icon>close</mat-icon>
             </button>
         </mat-form-field>
     
         <pre></pre>
 
-        <mat-form-field appearance="fill" [floatLabel]="passwordValue" class="input-form-field" hintLabel="Hasło musi mieć 6-30 znaków" [formGroup]="registrationFormGroup">
+        <mat-form-field appearance="fill" class="input-form-field" hintLabel="Hasło musi mieć 6-30 znaków" [formGroup]="registrationFormGroup">
             <mat-label>Hasło</mat-label>
             <input matInput type="text" placeholder="Hasło" #password required minlength="6" maxlength="30" formControlName="password">
-            <button mat-button *ngIf="passwordValue" matSuffix mat-icon-button aria-label="Clear" (click)="password.value=''">
+            <button mat-button matSuffix mat-icon-button aria-label="Clear" (click)="password.value=''">
                 <mat-icon>close</mat-icon>
             </button>
             <mat-hint align="end">{{password.value?.length || 0}}/30</mat-hint>
@@ -21,30 +21,30 @@
 
         <pre></pre>
 
-        <mat-form-field appearance="fill" [floatLabel]="firstNameValue" class="input-form-field" [formGroup]="registrationFormGroup">
+        <mat-form-field appearance="fill" class="input-form-field" [formGroup]="registrationFormGroup">
             <mat-label>Imię</mat-label>
             <input matInput type="text" placeholder="Imie" #firstName required formControlName="firstName">
-            <button mat-button *ngIf="firstNameValue" matSuffix mat-icon-button aria-label="Clear" (click)="firstName.value=''">
+            <button mat-button  matSuffix mat-icon-button aria-label="Clear" (click)="firstName.value=''">
                 <mat-icon>close</mat-icon>
             </button>
         </mat-form-field>
 
         <pre></pre>
 
-        <mat-form-field appearance="fill" [floatLabel]="lastNameValue" class="input-form-field" [formGroup]="registrationFormGroup">
+        <mat-form-field appearance="fill" class="input-form-field" [formGroup]="registrationFormGroup">
             <mat-label>Nazwisko</mat-label>
             <input matInput type="text" placeholder="Nazwisko" #lastName required formControlName="lastName">
-            <button mat-button *ngIf="lastNameValue" matSuffix mat-icon-button aria-label="Clear" (click)="lastName.value=''">
+            <button mat-button matSuffix mat-icon-button aria-label="Clear" (click)="lastName.value=''">
                 <mat-icon>close</mat-icon>
             </button>
         </mat-form-field>
 
         <pre></pre>
 
-        <mat-form-field appearance="fill" [floatLabel]="emailValue" class="input-form-field" [formGroup]="registrationFormGroup">
+        <mat-form-field appearance="fill" class="input-form-field" [formGroup]="registrationFormGroup">
             <mat-label>Email</mat-label>
             <input matInput type="email" [formControl]="emailFormControl" [errorStateMatcher]="matcher" placeholder="Email" #email required formControlName="email">
-            <button mat-button *ngIf="emailValue" matSuffix mat-icon-button aria-label="Clear" (click)="email.value=''">
+            <button mat-button matSuffix mat-icon-button aria-label="Clear" (click)="email.value=''">
                 <mat-icon>close</mat-icon>
             </button>
             <mat-error *ngIf="emailFormControl.hasError('email') && !emailFormControl.hasError('required')">
@@ -57,10 +57,10 @@
 
         <pre></pre>
 
-        <mat-form-field appearance="fill" [floatLabel]="phoneNumberValue" class="input-form-field" hintLabel="Numer telefonu musi mieć 9 cyfr" [formGroup]="registrationFormGroup">
+        <mat-form-field appearance="fill" class="input-form-field" hintLabel="Numer telefonu musi mieć 9 cyfr" [formGroup]="registrationFormGroup">
             <mat-label>Numer telefonu</mat-label>
             <input matInput type="text" placeholder="Numer telefonu" #phoneNumber required maxlength="9" formControlName="phoneNumber">
-            <button mat-button *ngIf="phoneNumberValue" matSuffix mat-icon-button aria-label="Clear" (click)="phoneNumber.value=''">
+            <button mat-button matSuffix mat-icon-button aria-label="Clear" (click)="phoneNumber.value=''">
                 <mat-icon>close</mat-icon>
             </button>
             <mat-hint align="end">{{phoneNumber.value?.length || 0}}/9</mat-hint>
