+++ src/app/registration/component/registration-info/registration-info.component.html
