+++ src/app/registration/component/registration-info/registration-info.component.ts
@@ -5,17 +5,17 @@ import { MatDialogRef, MAT_DIALOG_DATA } from '@angular/material/dialog';
   templateUrl: './registration-info.component.html',
   styleUrls: ['./registration-info.component.css']
 })
+
 export class RegistrationInfoComponent implements OnInit {
 
-  constructor(private dialogRef: MatDialogRef<RegistrationInfoComponent>,
-    @Inject(MAT_DIALOG_DATA) public data: any) { }
+  constructor(
+    private dialogRef: MatDialogRef<RegistrationInfoComponent>,
+    @Inject(MAT_DIALOG_DATA) public data: any
+  ) { }
 
-  onOkClick(): void{
+  onOkClick(): void {
     this.dialogRef.close();
   }
 
-  ngOnInit(): void {
-  }
-
-
+  ngOnInit(): void { }
 }
