+++ src/app/registration/model/RegistrationResponse.ts
