+++ src/app/registration/registration.module.ts
@@ -1,0 +1,49 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { RegistrationComponent } from './component/registation-user/registration.component';
+import { RegistrationInfoComponent } from './component/registration-info/registration-info.component';
+
+import { MatButtonModule } from '@angular/material/button';
+import { MatCardModule } from '@angular/material/card';
+import { MatFormFieldModule } from '@angular/material/form-field';
+import { MatInputModule } from '@angular/material/input';
+import { MatIconModule } from '@angular/material/icon';
+import { BrowserModule } from '@angular/platform-browser';
+import { HttpClientModule } from '@angular/common/http';
+import { AppRoutingModule } from '../app-routing.module';
+import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
+import { ReactiveFormsModule } from '@angular/forms';
+import { RouterModule } from '@angular/router';
+import { ErrorStateMatcher, ShowOnDirtyErrorStateMatcher } from '@angular/material/core';
+import { CookieService } from 'ngx-cookie-service';
+
+
+@NgModule({
+  declarations: [
+    RegistrationComponent,
+    RegistrationInfoComponent
+  ],
+  imports: [
+    CommonModule,
+    BrowserModule,
+    HttpClientModule,
+    AppRoutingModule,
+    BrowserAnimationsModule,
+    ReactiveFormsModule,
+    MatButtonModule,
+    MatCardModule,
+    MatFormFieldModule,
+    MatInputModule,
+    MatIconModule,
+    RouterModule.forRoot([
+      { path: 'registration', component: RegistrationComponent }
+    ]),
+  ],
+  providers: [
+    { 
+      provide: ErrorStateMatcher, useClass: ShowOnDirtyErrorStateMatcher
+    },
+    CookieService
+  ]
+})
+export class RegistrationModule { }
