+++ src/app/reservation/component/add-reservation/add-reservation.component.html
@@ -1,0 +1,107 @@
+<mat-card class="card">
+
+  <mat-card-title>
+    Dodaj rezerwację
+  </mat-card-title>
+
+  <mat-card-content>
+      <mat-label>Termin rezerwacji</mat-label>
+      <mat-form-field appearance="fill" class="input-form">
+        <mat-label>Wybierz termin:</mat-label>
+        <mat-date-range-input [rangePicker]="picker" [formGroup]="dateRange">
+          <input matStartDate formControlName="start" placeholder="Początek">
+          <input matEndDate formControlName="end" placeholder="Koniec">
+        </mat-date-range-input>
+        <mat-datepicker-toggle matSuffix [for]="picker"> </mat-datepicker-toggle>
+        <mat-date-range-picker #picker> </mat-date-range-picker>
+      </mat-form-field>
+  </mat-card-content>
+
+  <mat-card-content>
+    <mat-label>Adresat rezerwacji: </mat-label>
+    <mat-form-field appearance="fill" class="input-form">
+      <mat-select #User>
+        <mat-option *ngFor="let user of usersList" [value]="user">
+          {{user.firstName}} {{user.lastName}}
+        </mat-option>
+      </mat-select>
+    </mat-form-field>
+  </mat-card-content>
+
+  <mat-card-content>
+    <mat-label>Piętro: </mat-label>
+    <mat-form-field appearance="fill" class="input-form">
+      <mat-select #Floor (selectionChange)="getRooms(Floor.value, dateRange.value.start, dateRange.value.end)">
+        <mat-option *ngFor="let floor of floors" [value]="floor" >
+          {{floor}}
+        </mat-option>
+      </mat-select>
+    </mat-form-field>
+  </mat-card-content>
+
+  <mat-card-content>
+      <mat-label>Pokój:</mat-label>
+      <mat-form-field apperance="fill" class="input-form">
+        <mat-select #Room (selectionChange)="getWorksites(Room.value.name, dateRange.value.start, dateRange.value.end)">
+            <mat-option *ngFor="let room of roomsList" [value]="room" >
+              {{room.name}}
+            </mat-option>
+        </mat-select>
+      </mat-form-field>
+  </mat-card-content>
+
+  <mat-card-content>
+    <mat-label>Numer stanowiska w pokoju:</mat-label>
+    <mat-form-field apperance="fill" class="input-form">
+        <mat-select #WorkSite>
+            <mat-option *ngFor="let workSite of worksitesList" [value]="workSite">
+                {{workSite.worksiteInRoomId}}
+            </mat-option>
+        </mat-select>
+    </mat-form-field>
+</mat-card-content>
+
+<mat-card-content>
+  <mat-form-field class="chips-form" appearance="none">
+    <mat-chip-list #chipList aria-label="WorkSites" class="chips-form" required>
+      <mat-chip *ngFor="let worksite of reservatedWorkSites" 
+        [selectable]="selectableWorkSitesChips" 
+        [removable]="removable"
+        (removed)="removeChipFormList(worksite)" >
+
+         {{worksite.owner.firstName}} 
+         {{worksite.owner.lastName}} 
+         Piętro:{{worksite.floorNumber}} 
+         {{worksite.roomName}}
+         Numer miejsca:{{worksite.worksite.worksiteInRoomId}} 
+         Od: {{worksite.startDate}} do {{worksite.endDate}}
+
+        <mat-icon matChipRemove *ngIf="removable">cancel</mat-icon>
+      </mat-chip>
+    </mat-chip-list>
+  </mat-form-field>
+</mat-card-content>
+
+
+  <mat-card-actions>
+    <button mat-stroked-button class ="input-form-field" 
+      (click)="addChipToList(
+        User.value, 
+        Floor.value, 
+        Room.value.name, 
+        WorkSite.value, 
+        dateRange.value.start,
+        dateRange.value.end
+      )" >
+      Dodaj rezerwację
+    </button>
+  </mat-card-actions>
+
+  <mat-card-actions>
+    <button mat-stroked-button class ="input-form-field" 
+      (click)="postReservations()">
+      Zarezerwuj
+    </button>
+  </mat-card-actions>
+
+  </mat-card>
